user controller done

+++ controllers/userController.js
@@ -1,5 +1,16 @@
 const daoUser = require("../dao/userDao.js");
 
+isExistUser = id => {
+  let exist = false;
+  let user = daoUser.isExist(id);
+  if (user) {
+    exist = true;
+    return exist;
+  } else {
+    return exist;
+  }
+};
+
 exports.index = (req, res) => {
   daoUser.getAll((err, users) => {
     console.log("users", users);
@@ -11,8 +22,12 @@ exports.show = (req, res) => {
   let id = req.params.id;
 
   try {
-    let user = daoUser.getUser(id);
-    res.json({ user: user });
+    if (isExistUser(id)) {
+      let user = daoUser.getUser(id);
+      res.json({ user: user });
+    } else {
+      res.status(400).json("Usuario no encontrado");
+    }
   } catch (e) {
     console.log(e);
   }
@@ -22,8 +37,12 @@ exports.edit = (req, res) => {
   let id = req.params.id;
 
   try {
-    let user = daoUser.getUser(id);
-    res.json({ user: user });
+    if (isExistUser(id)) {
+      let user = daoUser.getUser(id);
+      res.json({ user: user });
+    } else {
+      res.status(400).json("Usuario no econtrado");
+    }
   } catch (e) {
     console.log(e);
   }
@@ -33,14 +52,18 @@ exports.update = (req, res) => {
   let { id, name, pass, active } = req.body;
   //agregar el updated at , fecha en que se lo modifica
   try {
-    user = {
-      id: id,
-      name: name,
-      pass: pass,
-      active: active
-    };
-    daoUser.updateUser(user);
-    res.json("Se edito correctamente el usuario");
+    if (isExistUser(id)) {
+      user = {
+        id: id,
+        name: name,
+        pass: pass,
+        active: active
+      };
+      daoUser.updateUser(user);
+      res.json("Se edito correctamente el usuario");
+    } else {
+      res.status(400).json("Usuario no encontrado");
+    }
   } catch (e) {
     console.log(e);
   }
@@ -50,8 +73,12 @@ exports.delete = (req, res) => {
   let id = req.params.id;
 
   try {
-    daoUser.deleteUser(id);
-    res.json("Se borro el usuario correctamente");
+    if (!isExistUser(id)) {
+      daoUser.deleteUser(id);
+      res.json("Se borro el usuario correctamente");
+    } else {
+      res.status(400).json("el usuario no existe o ya fue borrado");
+    }
   } catch (e) {
     console.log(e);
   }
